+++ app/src/main/java/ru/mkryuchkov/blogya/repository/PostPreviewRepository.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface PostPreviewRepository {
 
-    List<PostPreview> findAll(Integer limit, Integer offset);
+    List<PostPreview> findAll(Integer limit, Long offset);
 
-    List<PostPreview> findAllByTag(String tag, Integer limit, Integer offset);
+    List<PostPreview> findAllByTag(String tag, Integer limit, Long offset);
 
 }
